refactor: 테스트 코드 분리
시도 횟수와 관련된 테스트를 GameTotalCountTest로 분리하였습니다.
