+++ src/test/java/racingcar/util/VerificationUtilTest.java
@@ -39,24 +39,4 @@ class VerificationUtilTest {
         //then
         Assertions.assertThrows(IllegalArgumentException.class, () -> validateCarName(name));
     }
-
-    @Test
-    @DisplayName("시도 횟수가 정수인지 테스트")
-    public void 시도횟수_정수인지_테스트() {
-        // given
-        String attempt = "숫자";
-
-        //then
-        Assertions.assertThrows(IllegalArgumentException.class, () -> new GameTotalCount(attempt));
-    }
-
-    @Test
-    @DisplayName("시도 횟수가 음수인 경우 테스트")
-    public void 시도횟수_음수_테스트() {
-        // given
-        String attempt = "-3";
-
-        //then
-        Assertions.assertThrows(IllegalArgumentException.class, () -> new GameTotalCount(attempt));
-    }
 }
